+++ core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/datasource/MemoRemoteDataSourceModule.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.supabase.memo
+package app.kaito_dogi.mybrary.core.supabase.datasource
 
 import app.kaito_dogi.mybrary.core.data.datasource.MemoRemoteDataSource
 import dagger.Binds
