+++ core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/datasource/MyBookRemoteDataSourceModule.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.supabase.mybook
+package app.kaito_dogi.mybrary.core.supabase.datasource
 
 import app.kaito_dogi.mybrary.core.data.datasource.MyBookRemoteDataSource
 import dagger.Binds
