+++ core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/input/MemoInput.kt
@@ -1,5 +1,6 @@
-package app.kaito_dogi.mybrary.core.supabase.memo
+package app.kaito_dogi.mybrary.core.supabase.input
 
+import app.kaito_dogi.mybrary.core.data.command.PostMemoCommand
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -12,4 +13,11 @@ internal data class MemoInput(
   val startPage: Int?,
   @SerialName("end_page")
   val endPage: Int?,
+)
+
+internal fun PostMemoCommand.toMemoInput() = MemoInput(
+  myBookId = this.myBookId,
+  content = this.content,
+  startPage = this.startPage,
+  endPage = this.endPage,
 )
