--- core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/memo/MemoInputConvertor.kt
@@ -1,10 +1,0 @@
-package app.kaito_dogi.mybrary.core.supabase.memo
-
-import app.kaito_dogi.mybrary.core.data.command.PostMemoCommand
-
-internal fun PostMemoCommand.toMemoInput() = MemoInput(
-  myBookId = this.myBookId,
-  content = this.content,
-  startPage = this.startPage,
-  endPage = this.endPage,
-)
