--- core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/mybook/MyBookInputConvertor.kt
@@ -1,8 +1,0 @@
-package app.kaito_dogi.mybrary.core.supabase.mybook
-
-import app.kaito_dogi.mybrary.core.data.command.PostMyBookCommand
-
-internal fun PostMyBookCommand.toMyBookInput() = MyBookInput(
-  bookId = this.bookId,
-  userId = this.userId,
-)
