+++ core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/response/GenreResponse.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.supabase.book
+package app.kaito_dogi.mybrary.core.supabase.response
 
 import app.kaito_dogi.mybrary.core.data.dto.GenreDto
 import kotlinx.serialization.KSerializer
