+++ core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/response/MemoResponse.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.supabase.memo
+package app.kaito_dogi.mybrary.core.supabase.response
 
 import app.kaito_dogi.mybrary.core.data.dto.MemoDto
 import kotlinx.serialization.SerialName
